Extracted ForecastExternalMixin from ExternalMixin class in order to decouple the forecast and measurement command handling from trading external commands.
