+++ src/d3a/models/strategy/external_strategies/__init__.py
@@ -24,8 +24,7 @@ from typing import Dict, TYPE_CHECKING
 from d3a_interface.constants_limits import ConstSettings
 from d3a_interface.data_classes import Trade
 from d3a_interface.enums import SpotMarketTypeEnum
-from d3a_interface.utils import (convert_str_to_pendulum_in_dict,
-                                 str_to_pendulum_datetime)
+from d3a_interface.utils import str_to_pendulum_datetime
 from pendulum import DateTime
 
 import d3a.constants
@@ -485,8 +484,6 @@ class ExternalMixin:
             "delete_offer": self._delete_offer_aggregator,
             "list_offers": self._list_offers_aggregator,
             "device_info": self._device_info,
-            "set_energy_forecast": self._set_energy_forecast_aggregator,
-            "set_energy_measurement": self._set_energy_measurement_aggregator
         }
 
     def trigger_aggregator_commands(self, command: Dict) -> Dict:
@@ -527,108 +524,6 @@ class ExternalMixin:
                                   f"on area {self.device.name} with arguments {req.arguments}."
                                   "Market cycle already finished."})
         self.pending_requests = deque()
-
-    def _set_energy_forecast(self, payload: Dict) -> None:
-        """Callback for set_energy_forecast redis endpoint."""
-        self._set_device_energy_data(payload, "set_energy_forecast", "energy_forecast")
-
-    def _set_energy_measurement(self, payload: Dict) -> None:
-        """Callback for set_energy_measurement redis endpoint."""
-        self._set_device_energy_data(payload, "set_energy_measurement", "energy_measurement")
-
-    def _set_device_energy_data(self, payload: Dict, command_type: str,
-                                energy_data_arg_name: str) -> None:
-        transaction_id = self._get_transaction_id(payload)
-        energy_data_response_channel = f"{self.channel_prefix}/response/{command_type}"
-        arguments = json.loads(payload["data"])
-        if set(arguments.keys()) == {energy_data_arg_name, "transaction_id"}:
-            self.pending_requests.append(
-                IncomingRequest(command_type, arguments,
-                                energy_data_response_channel))
-        else:
-            logging.exception("Incorrect %s request: Payload %s", command_type, payload)
-            self.redis.publish_json(
-                energy_data_response_channel,
-                {"command": command_type,
-                 "error": f"Incorrect {command_type} request. "
-                          f"Available parameters: ({energy_data_arg_name}).",
-                 "transaction_id": transaction_id})
-
-    def _set_energy_forecast_impl(self, arguments: Dict, response_channel: str) -> None:
-        self._set_device_energy_data_impl(arguments, response_channel, "set_energy_forecast")
-
-    def _set_energy_measurement_impl(self, arguments: Dict, response_channel: str) -> None:
-        self._set_device_energy_data_impl(arguments, response_channel, "set_energy_measurement")
-
-    def _set_device_energy_data_impl(self, arguments: Dict, response_channel: str,
-                                     command_type: str) -> None:
-        """
-        Digest command from buffer and perform set_energy_forecast or set_energy_measurement.
-        Args:
-            arguments: Dictionary containing "energy_forecast/energy_measurement" that is a dict
-                containing a profile
-                key: time string, value: energy in kWh
-            response_channel: redis channel string where the response should be sent to
-        """
-        try:
-            self._set_device_energy_data_state(command_type, arguments)
-            response = {"command": command_type, "status": "ready",
-                        "transaction_id": arguments.get("transaction_id")}
-        except Exception:
-            error_message = (f"Error when handling _{command_type}_impl "
-                             f"on area {self.device.name}. Arguments: {arguments}")
-            logging.exception(error_message)
-            response = {"command": command_type, "status": "error",
-                        "error_message": error_message,
-                        "transaction_id": arguments.get("transaction_id")}
-        self.redis.publish_json(response_channel, response)
-
-    def _set_energy_forecast_aggregator(self, arguments: Dict) -> Dict:
-        """Callback for the set_energy_forecast endpoint when sent by aggregator."""
-        return self._set_device_energy_data_aggregator(arguments, "set_energy_forecast")
-
-    def _set_energy_measurement_aggregator(self, arguments: Dict) -> Dict:
-        """Callback for the set_energy_measurement endpoint when sent by aggregator."""
-        return self._set_device_energy_data_aggregator(arguments, "set_energy_measurement")
-
-    def _set_device_energy_data_aggregator(self, arguments: Dict, command_name: str) -> Dict:
-        try:
-            self._set_device_energy_data_state(command_name, arguments)
-            response = {
-                "command": command_name, "status": "ready",
-                command_name: arguments,
-                "area_uuid": self.device.uuid,
-                "transaction_id": arguments.get("transaction_id")}
-        except Exception:
-            logging.exception("Error when handling bid on area %s", self.device.name)
-            response = {
-                "command": command_name, "status": "error",
-                "area_uuid": self.device.uuid,
-                "error_message": f"Error when handling {command_name} "
-                                 f"on area {self.device.name} with arguments {arguments}.",
-                "transaction_id": arguments.get("transaction_id")}
-        return response
-
-    @staticmethod
-    def _validate_values_positive_in_profile(profile: Dict) -> None:
-        """Validate whether all values are positive in a profile."""
-        for time_str, energy in profile.items():
-            if energy < 0.0:
-                raise ValueError(f"Energy is not positive for time stamp {time_str}.")
-
-    def _set_device_energy_data_state(self, command_type: str, arguments: Dict) -> None:
-        """Update the state of the device with forecast and measurement data."""
-        if command_type == "set_energy_forecast":
-            self._validate_values_positive_in_profile(arguments["energy_forecast"])
-            self.energy_forecast_buffer.update(
-                convert_str_to_pendulum_in_dict(arguments["energy_forecast"]))
-        elif command_type == "set_energy_measurement":
-            self._validate_values_positive_in_profile(arguments["energy_measurement"])
-            self.energy_measurement_buffer.update(
-                convert_str_to_pendulum_in_dict(arguments["energy_measurement"]))
-        else:
-            assert False, f"Unsupported command {command_type}, available commands: " \
-                          "set_energy_forecast, set_energy_measurement"
 
     @property
     def market_info_dict(self) -> Dict:
