+++ src/d3a/models/strategy/external_strategies/pv.py
@@ -17,16 +17,16 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 import json
 import logging
-from collections import deque
 from typing import Dict
 
 from d3a_interface.constants_limits import ConstSettings
 from d3a_interface.data_classes import Offer
-from pendulum import duration, DateTime
+from pendulum import duration
 
 from d3a.d3a_core.util import get_market_maker_rate_from_config
 from d3a.models.strategy.external_strategies import (
     ExternalMixin, IncomingRequest, ExternalStrategyConnectionManager, default_market_info)
+from d3a.models.strategy.external_strategies.forecast_mixin import ForecastExternalMixin
 from d3a.models.strategy.predefined_pv import PVPredefinedStrategy, PVUserProfileStrategy
 from d3a.models.strategy.pv import PVStrategy
 
@@ -359,7 +359,7 @@ class PVPredefinedExternalStrategy(PVExternalMixin, PVPredefinedStrategy):
     """Strategy class for external PV devices with predefined profile."""
 
 
-class PVForecastExternalStrategy(PVPredefinedExternalStrategy):
+class PVForecastExternalStrategy(ForecastExternalMixin, PVPredefinedExternalStrategy):
     """
         Strategy responsible for reading forecast and measurement production data via hardware API
     """
@@ -387,62 +387,6 @@ class PVForecastExternalStrategy(PVPredefinedExternalStrategy):
                          energy_rate_decrease_per_update=energy_rate_decrease_per_update,
                          use_market_maker_rate=use_market_maker_rate)
 
-        # Buffers for energy forecast and measurement values,
-        # that have been sent by the d3a-api-client in the duration of one market slot
-        self.energy_forecast_buffer: Dict[DateTime, float] = {}
-        self.energy_measurement_buffer: Dict[DateTime, float] = {}
-
-    @property
-    def channel_dict(self) -> Dict:
-        """Extend channel_dict property with forecast related channels."""
-        return {**super().channel_dict,
-                f"{self.channel_prefix}/set_energy_forecast": self._set_energy_forecast,
-                f"{self.channel_prefix}/set_energy_measurement": self._set_energy_measurement}
-
-    def event_tick(self) -> None:
-        """Set the energy forecast using pending requests. Extends super implementation.
-
-        This method is triggered by the TICK event.
-        """
-        # Need to repeat the pending request parsing in order to handle energy forecasts
-        # from the MQTT subscriber (non-connected admin)
-        for req in self.pending_requests:
-            if req.request_type == "set_energy_forecast":
-                self._set_energy_forecast_impl(req.arguments, req.response_channel)
-            elif req.request_type == "set_energy_measurement":
-                self._set_energy_measurement_impl(req.arguments, req.response_channel)
-
-        self.pending_requests = deque(
-            req for req in self.pending_requests
-            if req.request_type not in ["set_energy_forecast", "set_energy_measurement"])
-
-        super().event_tick()
-
-    def _incoming_commands_callback_selection(self, req: IncomingRequest) -> None:
-        """Map commands to callbacks for forecast and measurement reading."""
-        if req.request_type == "set_energy_forecast":
-            self._set_energy_forecast_impl(req.arguments, req.response_channel)
-        elif req.request_type == "set_energy_measurement":
-            self._set_energy_measurement_impl(req.arguments, req.response_channel)
-        else:
-            super()._incoming_commands_callback_selection(req)
-
-    def event_market_cycle(self) -> None:
-        """Update forecast and measurement in state by reading from buffers."""
-        self.update_energy_forecast()
-        self.update_energy_measurement()
-        self._clear_energy_buffers()
-        super().event_market_cycle()
-
-    def _clear_energy_buffers(self) -> None:
-        """Clear forecast and measurement buffers."""
-        self.energy_forecast_buffer = {}
-        self.energy_measurement_buffer = {}
-
-    def event_activate_energy(self) -> None:
-        """This strategy does not need to initiate energy values as they are sent via the API."""
-        self._clear_energy_buffers()
-
     def update_energy_forecast(self) -> None:
         """Set energy forecast for future markets."""
         for slot_time, energy_kWh in self.energy_forecast_buffer.items():
@@ -454,11 +398,3 @@ class PVForecastExternalStrategy(PVPredefinedExternalStrategy):
         for slot_time, energy_kWh in self.energy_measurement_buffer.items():
             if slot_time < self.area.spot_market.time_slot:
                 self.state.set_energy_measurement_kWh(energy_kWh, slot_time)
-
-    def set_produced_energy_forecast_kWh_future_markets(self, reconfigure=False) -> None:
-        """
-        Setting produced energy for the next slot is already done by produced_energy_forecast_kWh
-        """
-
-    def _read_or_rotate_profiles(self, reconfigure=False) -> None:
-        """Overridden with empty implementation to disable reading profile from DB."""
